Adding 2d euclidean regression
To prevent infinite flipping, we pick a next event from a set of events happening more or less at the same time. Here, we re-order the events that are happening at time t=now. We now also fit a line through all points of the triangles that were selected. Iff these points are all on the same support line (are_residuals_near_zero), then infinite flipping could occur (so we proceed with re-ordering).